Change formatting of displayed default values in the commandline help:
The default values were displayed in square brackets and are now additionally
prefixed with the string 'default: ' within the brackets (as proposed in the
examples from the documentation of python's optparse module
(http://docs.python.org/library/optparse.html#optparse-generating-help).

The code now uses optparse's %default feature in the help strings for reduced maintenance effort.

On a side note: "Deprecated since version 2.7: The optparse module is deprecated
and will not be developed further; development will continue with the argparse
module."