Ignore encoding errors when parsing mails with Python 3.1
Python 3.1 doesn't have email.message_from_binary_file(), so when running
under this version of Python we need to use the old email.message_from_file()
function that takes an (already decoded) string rather than raw bytes. As we
used the built-in open() method (in text mode) to provide input to
message_from_file(), all errors were handled strictly, causing afew to abort
with a Traceback when it encountered a message that was encoded in a character
set different from the current locale.

Work around this by silently ignoring encoding errors. This means all
non-ASCII will be filtered out on many messages (which affects tagging), but
it's still better than breaking.

The real fix is to update Python to 3.2, but unfortunately Debian Squeeze
(stable) ships 3.1.3 only.

Signed-off-by: Sascha Silbe <sascha-pgp@silbe.org>