afew/MailMover.py: handle messages that exist in multiple places
Gmail in particular tends to put copies of a single message in
multiple maildirs.  notmuch groups these all into a single
message (with a single Message-ID) with multiple associated files.
This commit provides better support for moving mail files when
multiple copies of a message exist, e.g., when a copy already exists
in the destination maildir.

This commit also supports copying a message to multiple destinations.
A message that matches multiple queries will be copied to every
destination associated with those queries, rather than just the first.
Thus, it changes afew's movement behavior slightly.

(This change was a side-effect of implementing support for messages
that exist in multiple locations in the simplest way.  Moving a
message immediately from its source location without updating the
notmuch db will cause an error when subsequent queries match the
message but its associated file is no longer at the path notmuch
thinks it is.  To avoid this, we postpone deletions until all copies
associated with all rules have occurred.)