Implement archiving mails into folders
This adds an operation which can be used to file (archive) messages into
folders. The new MailArchiver class is actually quite similar to MailMover,
with the (important) addition that the destination folder name is expanded
for variables. Ccurrently only the message date is provided for expansion,
which allows to configure it like this:

  [MailArchiver]
  folders = INBOX
  INBOX = 'not tag:flagged and not tag:inbox and not tag:unread':Archives/{date.year}/{date.year}-{date.month:0>2d}

That will move seen and unflagged messages from INBOX into an archives
subfolder which contains the date and month of the message in its name. This
creates a folder structure like this:

  Archives/
    2015/
      2015-11/
      2015-12/
    2016/
      2016-01/
      2016-02/
      ...