If decode fails but email is base64, use regex to get base64 data
Some mailinglists just slam their signature at the end of the email
whether it's base64 encoded or not, leading to afew throwing exceptions
currently, because get_payload doesn't handle that case.

Instead we can try to grab the base64 part out of the message with a
regex and then decode that ourselves.

Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>