MailMover: Only run 'notmuch new' if any mails were actually moved
When running `afew --move`, the notmuch database is always updated, even if not
files were actually moved. Since this takes a non-trivial amount of time, the
time to run afew can actually be dominated by the database update, which is
annoying when putting afew into an automated 'check new mails' hook.

This adds a small optimisation where the database is only updated if any
messages were actually moved. This makes running `afew --move` substantially
faster in the common case were no messages need to be moved.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>