FolderNameFilter: support messages that exist in multiple locations
In handle_message, tag a message on the basis of the folder names
in *all* the filesystem paths for that message.

This improves support in particular for GMail, where the same message
may be stored in several different directories.  Notmuch knows to
group these together, and notmuch.message.Message.get_filenames()
returns all the paths for the given message.  This change simply runs
the previous code in a loop over these paths.