doc: fix path to source code
When being run in an environment where afew was not already
installed Sphinx would print the following warnings:

    docs/source/implementation.rst:11: WARNING: autodoc: failed to import class u'Database' from module u'afew.Database'; the following exception was raised:
    docs/source/implementation.rst:18: WARNING: autodoc: failed to import class u'Filter' from module u'afew.Filter'; the following exception was raised:
    docs/source/implementation.rst:24: WARNING: autodoc: failed to import module u'afew.Settings'; the following exception was raised:
    docs/source/implementation.rst:27: WARNING: autodoc: failed to import module u'afew.NotmuchSettings'; the following exception was raised:
    docs/source/implementation.rst:34: WARNING: autodoc: failed to import module u'afew.utils'; the following exception was raised:

Sphinx needs to be made aware of the location to the source
code so it can find the Python modules even when they aren't
installed already.  This is required when the documentation is
built in a clean environment such as when creating packages for
operating systems.

This fixes #184.