documentation: installation and usage in docs/
The instructions in README.rst duplicated some of the information in
the more complete docs, but not enough to really get started using
afew. It is better to avoid duplication and refer new users directly
to the full documentation.

The README.rst installation instructions were slightly more complete
and recent, so they were copied to the
docs/sources/installation.rst.

While at it, fix broken links in the rst markup.

Fixes: afewmail/afew/#23.