Split folders and preserve quoted names
If a folder in the `MailMover` section includes a
space (e.g. `[Gmail].Sent Mail`) there is currently no way to quote
the folder name such as to preserve the space in it. This change
replaces the standard `split()` function with one that understands
quotes. For the above example, the following config entry works:

        folders = "[Gmail].Sent Mail" other_folder

Fixes: #97

Signed-off-by: Nicolas Bock <nicolasbock@gmail.com>