Fix trailing ; in notmuch config breaking SentMailsFilter
The trailing semicolon resulted in the string split containing an empty string. When the notmuch
query is built later on this results in no messages being matched by the filter.

e.g.  in notmuch-config
```
primary_email=primary@domain.com
other_email=secondary@domain.com;tertiary@domain.com;
```

This fix ensures that the empty string is not added to the `my_addresses` set.

A potentially better fix would be to ensure that `notmuch_settings.get()` removes trailing ;