docs/quickstart: simplify and improve creation of notmuch hook
This drops the unnecessary touch and adds chmod in return.

Also instead of describing the intended content, add a printf command
to create the file resulting in a single code block.  The alternative is
a code block, then a file content block and after that another code
block for the chmod plus some surrounding prose.