MeFilter: drop advertizing me_tag option
The me_tag option does nothing that couldn't be done with the tags option
that works on all Filters.

Recommend using tags instead of me_tag in the documentation and drop the
custom handle_message. Add some code to make existing rules work as
expected.

As a side-effect this makes

	[MeFilter]
	tags = -something

work as expected. It used to do nothing, now it removes the something tag.