release.yml: run on published releases, invoke twine manually
This now runs on all published releases, and invokes
`python setup.py sdist` and `twine upload` manually - the
gh-action-pypi-publish was a bit too magic for my taste and didn't work
properly.

This runs only on testpypi on purpose - we'll point it to the real PyPi
once this has proven to work.