docs: Fix HeaderMatchingFilter example
The example should tag emails from Redmine with two tags: a fixed
"redmine" and a dynamic one based on the project name. However, the
configuration does something different: It tags *ALL* emails with the
"redmine" tag and additionally tags emails with the
"X-Redmine-Project" header with its content. The reason for tagging
all emails is that a missing header is considered having an empty
value and the pattern matches an empty string.

The fix is to ensure that the pattern does not match an empty string.