Fix the --enable-filters option
Filter objects must be initialized with the options found in the config file
but, in the case of `--enable-filters`, new objects were being initialized with
config options ignored.

This change makes use of the `configured_filter_chain` list of
already-initialized filter objects and trims it down by comparing filter class
names with the values passed in the command line for `--enable-filters`.

Closes: #40